+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/RawCustomResourceOperationsImpl.java
@@ -24,17 +24,24 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.fabric8.kubernetes.api.model.DeleteOptions;
+import io.fabric8.kubernetes.api.model.DeleteOptionsBuilder;
 import io.fabric8.kubernetes.api.model.DeletionPropagation;
 import io.fabric8.kubernetes.api.model.ListOptions;
 import io.fabric8.kubernetes.api.model.ListOptionsBuilder;
-import io.fabric8.kubernetes.api.model.Status;
 import io.fabric8.kubernetes.client.Config;
+import io.fabric8.kubernetes.client.GracePeriodConfigurable;
 import io.fabric8.kubernetes.client.KubernetesClientException;
+import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
+import io.fabric8.kubernetes.client.dsl.AnyNamespaceable;
+import io.fabric8.kubernetes.client.dsl.Deletable;
+import io.fabric8.kubernetes.client.dsl.Nameable;
+import io.fabric8.kubernetes.client.dsl.Namespaceable;
 import io.fabric8.kubernetes.client.dsl.base.CustomResourceDefinitionContext;
 import io.fabric8.kubernetes.client.dsl.base.OperationSupport;
 import io.fabric8.kubernetes.client.utils.HttpClientUtils;
@@ -57,19 +64,59 @@ import okhttp3.Response;
  * Right now it supports basic operations like GET, POST, PUT, DELETE.
  *
  */
-public class RawCustomResourceOperationsImpl extends OperationSupport {
+public class RawCustomResourceOperationsImpl extends OperationSupport implements Nameable<RawCustomResourceOperationsImpl>,
+  Namespaceable<RawCustomResourceOperationsImpl>,
+  AnyNamespaceable<RawCustomResourceOperationsImpl>,
+  GracePeriodConfigurable<RawCustomResourceOperationsImpl>,
+  PropagationPolicyConfigurable<RawCustomResourceOperationsImpl>,
+  Deletable {
   
   private static final String METADATA = "metadata";
   private static final String RESOURCE_VERSION = "resourceVersion";
   private final CustomResourceDefinitionContext customResourceDefinition;
   private final ObjectMapper objectMapper;
+  private final long gracePeriodInSeconds;
+  private final DeletionPropagation deletionPropagation;
 
   private enum HttpCallMethod { GET, POST, PUT, DELETE }
 
-  public RawCustomResourceOperationsImpl(OkHttpClient client, Config config, CustomResourceDefinitionContext customResourceDefinition) {
+  private RawCustomResourceOperationsImpl(OkHttpClient client, Config config, CustomResourceDefinitionContext crdContext, String namespace, String name, long gracePeriodInSeconds, DeletionPropagation deletionPropagation) {
     super(client, config);
-    this.customResourceDefinition = customResourceDefinition;
+    this.customResourceDefinition = crdContext;
     this.objectMapper = Serialization.jsonMapper();
+    this.namespace = namespace;
+    this.name = name;
+    this.gracePeriodInSeconds = gracePeriodInSeconds;
+    this.deletionPropagation = deletionPropagation;
+  }
+
+  public RawCustomResourceOperationsImpl(OkHttpClient client, Config config, CustomResourceDefinitionContext customResourceDefinition) {
+    this(client, config, customResourceDefinition, null, null, 0, null);
+  }
+
+  @Override
+  public RawCustomResourceOperationsImpl withName(String name) {
+    return new RawCustomResourceOperationsImpl(client, config, customResourceDefinition, namespace, name, gracePeriodInSeconds, deletionPropagation);
+  }
+
+  @Override
+  public RawCustomResourceOperationsImpl inNamespace(String namespace) {
+    return new RawCustomResourceOperationsImpl(client, config, customResourceDefinition, namespace, name, gracePeriodInSeconds, deletionPropagation);
+  }
+
+  @Override
+  public RawCustomResourceOperationsImpl inAnyNamespace() {
+    return new RawCustomResourceOperationsImpl(client, config, customResourceDefinition, null, name, gracePeriodInSeconds, deletionPropagation);
+  }
+
+  @Override
+  public RawCustomResourceOperationsImpl withGracePeriod(long gracePeriodSeconds) {
+    return new RawCustomResourceOperationsImpl(client, config, customResourceDefinition, namespace, name, gracePeriodSeconds, deletionPropagation);
+  }
+
+  @Override
+  public RawCustomResourceOperationsImpl withPropagationPolicy(DeletionPropagation propagationPolicy) {
+    return new RawCustomResourceOperationsImpl(client, config, customResourceDefinition, namespace, name, gracePeriodInSeconds, propagationPolicy);
   }
 
   /**
@@ -102,7 +149,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException exception in case of any network/read problems
    */
   public Map<String, Object> create(String objectAsString) throws IOException {
-    return validateAndSubmitRequest(null, null, objectAsString, HttpCallMethod.POST);
+    return validateAndSubmitRequest(objectAsString, HttpCallMethod.POST);
   }
 
   /**
@@ -113,7 +160,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of problems while reading HashMap
    */
   public Map<String, Object> create(Map<String, Object> object) throws IOException {
-    return validateAndSubmitRequest(null, null, objectMapper.writeValueAsString(object), HttpCallMethod.POST);
+    return validateAndSubmitRequest(objectMapper.writeValueAsString(object), HttpCallMethod.POST);
   }
 
   /**
@@ -125,7 +172,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of problems while reading JSON object
    */
   public Map<String, Object> create(String namespace, String objectAsString) throws IOException {
-    return validateAndSubmitRequest(namespace, null, objectAsString, HttpCallMethod.POST);
+    initOperationNamespaceAndName(namespace, null);
+    return validateAndSubmitRequest(objectAsString, HttpCallMethod.POST);
   }
 
   /**
@@ -136,7 +184,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of problems while reading file
    */
   public Map<String, Object> create(InputStream objectAsStream) throws IOException {
-    return validateAndSubmitRequest(null, null, IOHelpers.readFully(objectAsStream), HttpCallMethod.POST);
+    return validateAndSubmitRequest(IOHelpers.readFully(objectAsStream), HttpCallMethod.POST);
   }
 
   /**
@@ -148,7 +196,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of problems while reading file
    */
   public Map<String, Object> create(String namespace, InputStream objectAsStream) throws IOException {
-    return validateAndSubmitRequest(namespace, null, IOHelpers.readFully(objectAsStream), HttpCallMethod.POST);
+    initOperationNamespaceAndName(namespace, null);
+    return validateAndSubmitRequest(IOHelpers.readFully(objectAsStream), HttpCallMethod.POST);
   }
 
   /**
@@ -160,7 +209,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of problems faced while serializing HashMap
    */
   public Map<String, Object> create(String namespace, Map<String, Object> object) throws IOException {
-    return validateAndSubmitRequest(namespace, null, objectMapper.writeValueAsString(object), HttpCallMethod.POST);
+    initOperationNamespaceAndName(namespace, null);
+    return validateAndSubmitRequest(objectMapper.writeValueAsString(object), HttpCallMethod.POST);
   }
 
   /**
@@ -172,7 +222,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serializiation failures or failures from Kuberntes API
    */
   public Map<String, Object> createOrReplace(String objectAsString) throws IOException {
-    return createOrReplaceObject(null, load(objectAsString));
+    return createOrReplaceObject(load(objectAsString));
   }
 
   /**
@@ -183,7 +233,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> createOrReplace(Map<String, Object> customResourceObject) throws IOException {
-    return createOrReplaceObject(null, customResourceObject);
+    return createOrReplaceObject(customResourceObject);
   }
 
   /**
@@ -194,7 +244,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> createOrReplace(InputStream inputStream) throws IOException {
-    return createOrReplaceObject(null, load(inputStream));
+    return createOrReplaceObject(load(inputStream));
   }
 
   /**
@@ -206,7 +256,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> createOrReplace(String namespace, String objectAsString) throws IOException {
-    return createOrReplaceObject(namespace, load(objectAsString));
+    initOperationNamespaceAndName(namespace, null);
+    return createOrReplaceObject(load(objectAsString));
   }
 
   /**
@@ -218,7 +269,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> createOrReplace(String namespace, Map<String, Object> customResourceObject) throws IOException {
-    return createOrReplaceObject(namespace, customResourceObject);
+    initOperationNamespaceAndName(namespace, null);
+    return createOrReplaceObject(customResourceObject);
   }
 
   /**
@@ -230,7 +282,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> createOrReplace(String namespace, InputStream objectAsStream) throws IOException {
-    return createOrReplaceObject(namespace, load(objectAsStream));
+    initOperationNamespaceAndName(namespace, null);
+    return createOrReplaceObject(load(objectAsStream));
   }
 
   /**
@@ -242,7 +295,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> edit(String name, Map<String, Object> object) throws IOException {
-    return validateAndSubmitRequest(null, name, objectMapper.writeValueAsString(object), HttpCallMethod.PUT);
+    initOperationNamespaceAndName(null, name);
+    return validateAndSubmitRequest(objectMapper.writeValueAsString(object), HttpCallMethod.PUT);
   }
 
   /**
@@ -254,7 +308,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> edit(String name, String objectAsString) throws IOException {
-    return validateAndSubmitRequest(null, name, objectAsString, HttpCallMethod.PUT);
+    initOperationNamespaceAndName(null, name);
+    return validateAndSubmitRequest(objectAsString, HttpCallMethod.PUT);
   }
 
   /**
@@ -267,8 +322,9 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> edit(String namespace, String name, Map<String, Object> object) throws IOException {
+    initOperationNamespaceAndName(namespace, name);
     object = appendResourceVersionInObject(namespace, name, object);
-    return validateAndSubmitRequest(namespace, name, objectMapper.writeValueAsString(object), HttpCallMethod.PUT);
+    return validateAndSubmitRequest(objectMapper.writeValueAsString(object), HttpCallMethod.PUT);
   }
 
   /**
@@ -281,10 +337,11 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/serialization failures or failures from Kubernetes API
    */
   public Map<String, Object> edit(String namespace, String name, String objectAsString) throws IOException {
+    initOperationNamespaceAndName(namespace, name);
     // Append resourceVersion in object metadata in order to
     // avoid : https://github.com/fabric8io/kubernetes-client/issues/1724
     objectAsString = appendResourceVersionInObject(namespace, name, objectAsString);
-    return validateAndSubmitRequest(namespace, name, objectAsString, HttpCallMethod.PUT);
+    return validateAndSubmitRequest(objectAsString, HttpCallMethod.PUT);
   }
 
   /**
@@ -297,7 +354,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case any failure to parse Map
    */
   public Map<String, Object> updateStatus(String name, Map<String, Object> objectAsMap) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(null, name, null) + "/status", objectMapper.writeValueAsString(objectAsMap), HttpCallMethod.PUT);
+    initOperationNamespaceAndName(null, name);
+    return validateAndSubmitRequest(fetchUrl(null) + "/status", objectMapper.writeValueAsString(objectAsMap), HttpCallMethod.PUT);
   }
 
   /**
@@ -310,7 +368,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case any failure to parse Map
    */
   public Map<String, Object> updateStatus(String name, String objectAsJsonString) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(null, name, null) + "/status", objectAsJsonString, HttpCallMethod.PUT);
+    initOperationNamespaceAndName(null, name);
+    return validateAndSubmitRequest(fetchUrl(null) + "/status", objectAsJsonString, HttpCallMethod.PUT);
   }
 
   /**
@@ -324,7 +383,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case any failure to parse Map
    */
   public Map<String, Object> updateStatus(String namespace, String name, Map<String, Object> objectAsMap) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(namespace, name, null) + "/status", objectMapper.writeValueAsString(objectAsMap), HttpCallMethod.PUT);
+    initOperationNamespaceAndName(namespace, name);
+    return validateAndSubmitRequest(fetchUrl(null) + "/status", objectMapper.writeValueAsString(objectAsMap), HttpCallMethod.PUT);
   }
 
   /**
@@ -337,7 +397,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case any failure to parse Map
    */
   public Map<String, Object> updateStatus(String name, InputStream objectAsStream) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(null, name, null) + "/status", IOHelpers.readFully(objectAsStream), HttpCallMethod.PUT);
+    initOperationNamespaceAndName(null, name);
+    return validateAndSubmitRequest(fetchUrl(null) + "/status", IOHelpers.readFully(objectAsStream), HttpCallMethod.PUT);
   }
 
   /**
@@ -351,7 +412,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case any failure to parse Map
    */
   public Map<String, Object> updateStatus(String namespace, String name, InputStream objectAsStream) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(namespace, name, null) + "/status", IOHelpers.readFully(objectAsStream), HttpCallMethod.PUT);
+    initOperationNamespaceAndName(namespace, name);
+    return validateAndSubmitRequest(fetchUrl(null) + "/status", IOHelpers.readFully(objectAsStream), HttpCallMethod.PUT);
   }
 
   /**
@@ -365,7 +427,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case any failure to parse Map
    */
   public Map<String, Object> updateStatus(String namespace, String name, String objectAsJsonString) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(namespace, name, null) + "/status", objectAsJsonString, HttpCallMethod.PUT);
+    initOperationNamespaceAndName(namespace, name);
+    return validateAndSubmitRequest(fetchUrl(null) + "/status", objectAsJsonString, HttpCallMethod.PUT);
   }
 
   /**
@@ -375,7 +438,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @return Object as HashMap
    */
   public Map<String, Object> get(String name) {
-    return makeCall(fetchUrl(null, name, null), null, HttpCallMethod.GET);
+    initOperationNamespaceAndName(null, name);
+    return makeCall(fetchUrl(null), null, HttpCallMethod.GET);
   }
 
   /**
@@ -386,7 +450,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @return Object as HashMap
    */
   public Map<String, Object> get(String namespace, String name) {
-      return makeCall(fetchUrl(namespace, name, null), null, HttpCallMethod.GET);
+    initOperationNamespaceAndName(namespace, name);
+    return makeCall(fetchUrl(null), null, HttpCallMethod.GET);
   }
 
   /**
@@ -395,7 +460,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @return list of custom resources as HashMap
    */
   public Map<String, Object> list() {
-    return makeCall(fetchUrl(null, null, null), null, HttpCallMethod.GET);
+    return makeCall(fetchUrl(null), null, HttpCallMethod.GET);
   }
 
   /**
@@ -405,7 +470,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @return list of custom resources as HashMap
    */
   public Map<String, Object> list(String namespace) {
-    return makeCall(fetchUrl(namespace, null, null), null, HttpCallMethod.GET);
+    initOperationNamespaceAndName(namespace, null);
+    return makeCall(fetchUrl(null), null, HttpCallMethod.GET);
   }
 
   /**
@@ -416,43 +482,80 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @return list of custom resources as HashMap
    */
   public Map<String, Object> list(String namespace, Map<String, String> labels) {
-    return makeCall(fetchUrl(namespace, null, labels), null, HttpCallMethod.GET);
+    initOperationNamespaceAndName(namespace, null);
+    return makeCall(fetchUrl(labels), null, HttpCallMethod.GET);
   }
 
   /**
-   * Delete all custom resources in a specific namespace
+   * Delete all Namespaced Scoped Custom Resources in a specified namespace
+   * <b>OR</b>
+   * Delete a Cluster Scoped Custom Resource with specified name
    *
-   * @param namespace desired namespace
-   * @return a boolean value whether item was deleted or item didn't exist in server
+   * <p>
+   * Note: This method behaves differently based on the scope of CRD:
+   *
+   * If specified CRD is of Namespaced scope, this method would delete all Custom Resources in namespace which is
+   * specified as parameter.
+   * If specified CRD is of Cluster scope, this method would delete a Custom Resource matching the name which is
+   * specified as parameter
+   * </p>
+   *
+   * @param namespaceOrName desired namespace(if CRD is Namespaced) or name(If CRD is Cluster)
+   * @return deleted objects as HashMap
    */
-  public boolean delete(String namespace) {
-    return handleDelete(namespace, null, null);
+  public boolean delete(String namespaceOrName) {
+    setNamespaceOrNameBasedOnScope(namespaceOrName);
+    return handleDelete(null);
   }
 
   /**
-   * Delete all custom resources in a specific namespace
+   * Delete all Namespaced Scoped Custom Resources in a specified namespace
+   * <b>OR</b>
+   * Delete a Cluster Scoped Custom Resource with specified name
    *
-   * @param namespace desired namespace
+   * <p>
+   * Note: This method behaves differently based on the scope of CRD:
+   *
+   * If specified CRD is of Namespaced scope, this method would delete all Custom Resources in namespace which is
+   * specified as parameter.
+   * If specified CRD is of Cluster scope, this method would delete a Custom Resource matching the name which is
+   * specified as parameter
+   * </p>
+   *
+   * @param namespaceOrName desired namespace(If CRD is Namespaced) or name(If CRD is Cluster)
    * @param cascading whether dependent object need to be orphaned or not.  If true/false, the "orphan"
    *                   finalizer will be added to/removed from the object's finalizers list.
    * @return a boolean value whether item was deleted or item didn't exist in server
    * @throws IOException in case of any network/parsing exception
    */
-  public boolean delete(String namespace, boolean cascading) throws IOException {
-    return handleDelete(namespace, null,  objectMapper.writeValueAsString(fetchDeleteOptions(cascading, null)));
+  public boolean delete(String namespaceOrName, boolean cascading) throws IOException {
+    setNamespaceOrNameBasedOnScope(namespaceOrName);
+    return handleDelete(objectMapper.writeValueAsString(fetchDeleteOptions(cascading, null)));
   }
 
   /**
-   * Delete all custom resources in a specific namespace
+   * Delete all Namespaced Scoped Custom Resources in a specified namespace
+   * <b>OR</b>
+   * Delete a Cluster Scoped Custom Resource with specified name
    *
-   * @param namespace desired namespace
+   * <p>
+   * Note: This method behaves differently based on the scope of CRD:
+   *
+   * If specified CRD is of Namespaced scope, this method would delete all Custom Resources in namespace which is
+   * specified as parameter.
+   * If specified CRD is of Cluster scope, this method would delete a Custom Resource matching the name which is
+   * specified as parameter
+   * </p>
+   *
+   * @param namespaceOrName desired namespace(If CRD is Namespaced) or name(If CRD is Cluster)
    * @param deleteOptions object provided by Kubernetes API for more fine grained control over deletion.
    *                       For more information please see https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.16/#deleteoptions-v1-meta
    * @return a boolean value whether item was deleted or item didn't exist in server
    * @throws IOException in case of any network/object parse problems
    */
-  public boolean delete(String namespace, DeleteOptions deleteOptions) throws IOException {
-    return handleDelete(namespace, null, objectMapper.writeValueAsString(deleteOptions));
+  public boolean delete(String namespaceOrName, DeleteOptions deleteOptions) throws IOException {
+    setNamespaceOrNameBasedOnScope(namespaceOrName);
+    return handleDelete(objectMapper.writeValueAsString(deleteOptions));
   }
 
   /**
@@ -464,7 +567,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of any network/object parse problems
    */
   public boolean delete(String namespace, String name) throws IOException {
-    return handleDelete(namespace, name, objectMapper.writeValueAsString(fetchDeleteOptions(false, DeletionPropagation.BACKGROUND.toString())));
+    initOperationNamespaceAndName(namespace, name);
+    return handleDelete(objectMapper.writeValueAsString(fetchDeleteOptions(false, DeletionPropagation.BACKGROUND.toString())));
   }
 
   /**
@@ -478,7 +582,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException exception related to network/object parsing
    */
   public boolean delete(String namespace, String name, boolean cascading) throws IOException {
-    return handleDelete(namespace, name, objectMapper.writeValueAsString(fetchDeleteOptions(cascading, null)));
+    initOperationNamespaceAndName(namespace, name);
+    return handleDelete(objectMapper.writeValueAsString(fetchDeleteOptions(cascading, null)));
   }
 
   /**
@@ -497,7 +602,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network/object parse exception
    */
   public boolean delete(String namespace, String name, String propagationPolicy) throws IOException {
-    return handleDelete(namespace, name, objectMapper.writeValueAsString(fetchDeleteOptions(false, propagationPolicy)));
+    initOperationNamespaceAndName(namespace, name);
+    return handleDelete(objectMapper.writeValueAsString(fetchDeleteOptions(false, propagationPolicy)));
   }
 
   /**
@@ -511,7 +617,17 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of any network/object parse exception
    */
   public boolean delete(String namespace, String name, DeleteOptions deleteOptions) throws IOException {
-    return handleDelete(namespace, name, objectMapper.writeValueAsString(deleteOptions));
+    initOperationNamespaceAndName(namespace, name);
+    return handleDelete(objectMapper.writeValueAsString(deleteOptions));
+  }
+
+  @Override
+  public Boolean delete() {
+    try {
+      return handleDelete(objectMapper.writeValueAsString(fetchDeleteOptions(false, null)));
+    } catch (JsonProcessingException jpe) {
+      throw KubernetesClientException.launderThrowable(jpe);
+    }
   }
 
   /**
@@ -598,11 +714,12 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
    * @throws IOException in case of network error
    */
   public Watch watch(String namespace, String name, Map<String, String> labels, ListOptions options, Watcher<String> watcher) throws IOException {
+    initOperationNamespaceAndName(namespace, name);
     if (options == null) {
       options = new ListOptions();
     }
     options.setWatch(true);
-    HttpUrl.Builder watchUrlBuilder = fetchWatchUrl(namespace, name, labels, options);
+    HttpUrl.Builder watchUrlBuilder = fetchWatchUrl(labels, options);
 
     OkHttpClient.Builder clonedClientBuilder = client.newBuilder();
       clonedClientBuilder.readTimeout(getConfig() != null ?
@@ -650,7 +767,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
 
   }
 
-  private Map<String, Object> createOrReplaceObject(String namespace, Map<String, Object> objectAsMap) throws IOException {
+  private Map<String, Object> createOrReplaceObject(Map<String, Object> objectAsMap) throws IOException {
     Map<String, Object> metadata = (Map<String, Object>) objectAsMap.get(METADATA);
     if (metadata == null) {
       throw KubernetesClientException.launderThrowable(new IllegalStateException("Invalid object provided -- metadata is required."));
@@ -663,8 +780,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
     metadata.remove(RESOURCE_VERSION);
 
     try {
-      if(namespace != null) {
-        ret = create(namespace, objectAsMap);
+      if(this.namespace != null) {
+        ret = create(this.namespace, objectAsMap);
       } else {
         ret = create(objectAsMap);
       }
@@ -678,9 +795,9 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
         if (originalResourceVersion != null) {
           metadata.put(RESOURCE_VERSION, originalResourceVersion);
         }
-        String name = (String) metadata.get("name");
-        ret = namespace != null ?
-          edit(namespace, name, objectAsMap) : edit(name, objectAsMap);
+        String nameFromObject = (String) metadata.get("name");
+        ret = this.namespace != null ?
+          edit(this.namespace, nameFromObject, objectAsMap) : edit(nameFromObject, objectAsMap);
       } catch (NullPointerException nullPointerException) {
         throw KubernetesClientException.launderThrowable(new IllegalStateException("Invalid object provided -- metadata.name is required."));
       }
@@ -706,27 +823,38 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
     return retVal;
   }
 
-  protected HttpUrl.Builder fetchWatchUrl(String namespace, String name, Map<String, String> labels, ListOptions options) throws MalformedURLException {
-    String resourceUrl = fetchUrl(namespace, null, labels);
+  protected HttpUrl.Builder fetchWatchUrl(Map<String, String> labels, ListOptions options) throws MalformedURLException {
+    String configuredName = this.name;
+    // Get Url without resource name configured
+    this.name = null;
+    String resourceUrl = fetchUrl(labels);
     if (resourceUrl.endsWith("/")) {
       resourceUrl = resourceUrl.substring(0, resourceUrl.length() - 1);
     }
     URL url = new URL(resourceUrl);
     HttpUrl.Builder httpUrlBuilder = HttpUrl.get(url).newBuilder();
 
-    if (name != null) {
-      httpUrlBuilder.addQueryParameter("fieldSelector", "metadata.name=" + name);
+    if (configuredName != null) {
+      httpUrlBuilder.addQueryParameter("fieldSelector", "metadata.name=" + configuredName);
     }
 
     HttpClientUtils.appendListOptionParams(httpUrlBuilder, options);
     return httpUrlBuilder;
   }
 
-  private String fetchUrl(String namespace, String name, Map<String, String> labels) {
+  private String fetchUrl(Map<String, String> labels) {
     if (config.getMasterUrl() == null) {
       return null;
     }
 
+    HttpUrl.Builder urlBuilder = HttpUrl.get(getNamespacedUrl(namespace, name)).newBuilder();
+    if (labels != null) {
+      urlBuilder.addQueryParameter("labelSelector", getLabelsQueryParam(labels));
+    }
+    return urlBuilder.toString();
+  }
+
+  private String getNamespacedUrl(String namespace, String name) {
     StringBuilder urlBuilder = new StringBuilder(config.getMasterUrl());
 
     urlBuilder.append(config.getMasterUrl().endsWith("/") ? "" : "/");
@@ -743,9 +871,6 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
     if (name != null) {
       urlBuilder.append("/").append(name);
     }
-    if(labels != null) {
-      urlBuilder.append("?labelSelector").append("=").append(getLabelsQueryParam(labels));
-    }
     return urlBuilder.toString();
   }
 
@@ -755,7 +880,7 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
       if(labelQueryBuilder.length() > 0) {
         labelQueryBuilder.append(",");
       }
-      labelQueryBuilder.append(entry.getKey()).append(Utils.toUrlEncoded("=")).append(entry.getValue());
+      labelQueryBuilder.append(entry.getKey()).append("=").append(entry.getValue());
     }
     return labelQueryBuilder.toString();
   }
@@ -788,8 +913,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
     return objectMapper.readValue(response.body().string(), HashMap.class);
   }
 
-  private boolean handleDelete(String namespace, String name, String requestBody) {
-    Map<String, Object> response = makeCall(fetchUrl(namespace, name, null), requestBody, HttpCallMethod.DELETE, false);
+  private boolean handleDelete(String requestBody) {
+    Map<String, Object> response = makeCall(fetchUrl(null), requestBody, HttpCallMethod.DELETE, false);
 
     // In most cases Status object is sent on deletion, but when deprecated DeleteOptions.orphanDependents
     // is used; object which is being deleted is sent
@@ -799,8 +924,8 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
     return true;
   }
 
-  private Map<String, Object> validateAndSubmitRequest(String namespace, String name, String objectAsString, HttpCallMethod httpCallMethod) throws IOException {
-    return validateAndSubmitRequest(fetchUrl(namespace, name, null), objectAsString, httpCallMethod);
+  private Map<String, Object> validateAndSubmitRequest(String objectAsString, HttpCallMethod httpCallMethod) throws IOException {
+    return validateAndSubmitRequest(fetchUrl(null), objectAsString, httpCallMethod);
   }
 
   private Map<String, Object> validateAndSubmitRequest(String resourceUrl, String objectAsString, HttpCallMethod httpCallMethod) throws IOException {
@@ -855,12 +980,38 @@ public class RawCustomResourceOperationsImpl extends OperationSupport {
   }
 
   private DeleteOptions fetchDeleteOptions(boolean cascading, String propagationPolicy) {
-    DeleteOptions deleteOptions = new DeleteOptions();
-    if (propagationPolicy != null) {
-      deleteOptions.setPropagationPolicy(propagationPolicy);
+    DeleteOptionsBuilder deleteOptionsBuilder = new DeleteOptionsBuilder();
+    String resolvedPropagationPolicy = resolvePropagationPolicy(propagationPolicy);
+    if (resolvedPropagationPolicy != null) {
+      deleteOptionsBuilder.withPropagationPolicy(resolvedPropagationPolicy);
     } else {
-      deleteOptions.setOrphanDependents(!cascading);
+      deleteOptionsBuilder.withOrphanDependents(!cascading);
     }
-    return deleteOptions;
+    if (this.gracePeriodInSeconds > 0) {
+      deleteOptionsBuilder.withGracePeriodSeconds(this.gracePeriodInSeconds);
+    }
+    return deleteOptionsBuilder.build();
+  }
+
+  private String resolvePropagationPolicy(String propagationPolicyProvided) {
+    if (propagationPolicyProvided != null) {
+      return propagationPolicyProvided;
+    } else if (this.deletionPropagation != null) {
+      return this.deletionPropagation.toString();
+    }
+    return null;
+  }
+
+  private void initOperationNamespaceAndName(String namespace, String name) {
+    this.namespace = namespace;
+    this.name = name;
+  }
+
+  private void setNamespaceOrNameBasedOnScope(String namespaceOrName) {
+    if (customResourceDefinition.getScope().equals("Namespaced")) {
+      this.namespace = namespaceOrName;
+    } else {
+      this.name = namespaceOrName;
+    }
   }
 }